requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Sockets==0.2.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Naked==0.1.31
peewee==3.9.5
pkg-resources==0.0.0
pycrypto==2.6.1
PyMySQL==0.9.3
PyYAML==5.1
requests==2.22.0
shellescape==3.4.1
six==1.12.0
urllib3==1.25.2
Werkzeug==0.15.4
